requirements.txt
streamlit==1.32.0
pandas==2.1.4
pillow==10.1.0
requests==2.31.0
